+++ components/projectPage/ProjectContent.tsx
